Fixed problem involving initial port not being set
The way that the Port was originally being managed would not set an init
ial port unless the -p command was initialized.  This was due to PORT be
ing stored as an int rather than a character… and then doing the convers
ion based on the input of the -p option.  Due to this, the PORT would be
 non existent if the -p option was not utilized.